requirements.txt
amqp==2.5.2
anyjson==0.3.3
appdirs==1.4.3
apturl==0.5.2
asgiref==3.4.1
astroid==2.3.3
bcrypt==3.1.7
beautifulsoup4==4.10.0
billiard==3.6.1.0
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
celery==4.2.1
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.9
djangorestframework==3.12.4
duplicity==0.8.12.0
entrypoints==0.3
ephem==3.7.7.1
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
isort==4.3.4
keyring==18.0.1
kombu==4.6.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
logilab-common==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
mininet==2.3.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
ovs==2.13.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.4
pycairo==1.16.2
pycodestyle==2.5.0
pycups==1.9.73
pyflakes==2.1.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-memcached==1.58
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sortedcontainers==2.1.0
soupsieve==2.3.1
sqlparse==0.4.2
ssh-import-id==5.10
supervisor==4.1.0
systemd-python==234
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vine==1.3.0
virtualenv==20.0.17
wadllib==1.3.3
wrapt==1.11.2
xkit==0.0.0
zipp==1.0.0
